requirements.txt
jinja2
beautifulsoup4
lxml
requests
python-dateutil
flask
